n-nope, we shall return 0
git-svn-id: http://svn.miranda-ng.org/main/trunk@1580 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ protocols/IcqOscarJ/fam_01service.cpp
@@ -684,7 +684,7 @@ void CIcqProto::setUserInfo()
 	wAdditionalData += 16;
 #endif
 
-	wAdditionalData += CustomCapList.size() * 16;
+	wAdditionalData += (WORD)CustomCapList.size() * 16;
 
 	//MIM/PackName
 	bool bHasPackName = false;
